Added reviews.json file/Updated movie-database
Added reviews.json file with 5 sample reviews. Updated movie-database.js file to support addition of basic and full reviews, and getting particular reviews as well as getting reviews for a particular movie. Changed the movieRating function to updateMovieRating so that whenever a review (basic or full) is added, the rating of a particular movie is updated as well. Also implemented a similarMovie function by checking if any 2 movies have more than 2 common genres. Will probably need to clear that up a bit more but it works.